src/graph_ident.c: ident_data_to_json: Average data points ...
... to send between 400 and 800 values. This reduces the time taken to
receive the values sent by the "instance_data_json" action considerably.
We should, however, do this in the data provider function or possibly in
the "instance_data_json" action. Also supporting other CFs would be a
wise move and while we're scanning the data we might compute total
average, minimum and maximum as well.